again PDE
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2672 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/pde/application/PDESolver.cpp
@@ -255,16 +255,16 @@ void PDESolver::coarsenInitialGridSurplus(sg::base::DataVector& alpha, double dT
 	}
 }
 
-void PDESolver::printGrid(sg::base::DataVector& alpha, size_t PointesPerDimension, std::string tfilename) const
+void PDESolver::printGrid(sg::base::DataVector& alpha, double PointesPerDimension, std::string tfilename) const
 {
 	sg::base::GridPrinter myPrinter(*this->myGrid);
-	myPrinter.printGrid(alpha, tfilename, PointesPerDimension);
+	myPrinter.printGrid(alpha, tfilename, static_cast<size_t>(PointesPerDimension));
 }
 
-void PDESolver::printGridDomain(sg::base::DataVector& alpha, size_t PointesPerDimension, sg::base::BoundingBox& GridArea, std::string tfilename) const
+void PDESolver::printGridDomain(sg::base::DataVector& alpha, double PointesPerDimension, sg::base::BoundingBox& GridArea, std::string tfilename) const
 {
 	sg::base::GridPrinter myPrinter(*this->myGrid);
-	myPrinter.printGridDomain(alpha, tfilename, GridArea, PointesPerDimension);
+	myPrinter.printGridDomain(alpha, tfilename, GridArea, static_cast<size_t>(PointesPerDimension));
 }
 
 void PDESolver::printSparseGrid(sg::base::DataVector& alpha, std::string tfilename, bool bSurplus) const
